+++ docsig/_function.py
@@ -160,10 +160,16 @@ class Function:
     """Represents a function with signature and docstring parameters.
 
     :param node: Function's abstract syntax tree.
+    :param doc_node: Docstring node if other than this function.
     :param method: Boolean for whether function is a class method.
     """
 
-    def __init__(self, node: _ast.FunctionDef, method: bool = False) -> None:
+    def __init__(
+        self,
+        node: _ast.FunctionDef,
+        doc_node: _ast.Const | None = None,
+        method: bool = False,
+    ) -> None:
         self._name = node.name
         self._lineno = node.lineno or 0
         self._isproperty = False
@@ -176,7 +182,7 @@ class Function:
         self._signature = Signature(
             node.args, node.returns, method=method, prop=self._isproperty
         )
-        self._docstring = Docstring(node.doc_node)
+        self._docstring = Docstring(doc_node or node.doc_node)
 
     @property
     def name(self) -> str: